refactor: Исправляет замечания наставника
